Fix Sphinx warnings.
Most of them were due to invisible-code-blocks. As per
http://svn.zope.org/manuel/trunk/src/manuel/index.txt:

The "invisible-code-block" directive will work with either one or two
colons.  The reason is that reST processers (like docutils and Sphinx) will
generate an error for unrecognized directives (like invisible-code-block).
Therefore you can use a single colon and the line will be interpreted as a
comment instead.

For this reason we are now using a single colon after the
invisible-code-block directive.