Handle automatic transitions
If content.transition is called with a to_state parameter,
a list of transitions that make possible to move from the object original
state to the final desired state is compiled with ``_wf_transitions_for``.

If that list of transitions contain any automatic transition and is triggered
the iteration over that list of transitions will be broken as the
automatic transition will have already happened.

This commit handles this case.